pyproject.toml
[tool.poetry]
name = "real-estate"
version = "0.1.0"
description = ""
authors = ["max-kovalenko <max.kovalenko.workshop@gmail.com>"]
readme = "README.md"
packages = [{include = "real_estate"}]

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.1.4"
numpy = "^1.26.3"
scikit-learn = "^1.3.2"
lightgbm = "^4.2.0"
hydra-core = "^1.3.2"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[tool.black]
line-length = 90
target-version = ["py310"]

[tool.isort]
src_paths = ["ds_project"]
profile = "black"
line_length = 90
lines_after_imports = 2

[tool.flake8]
ignore = ['E203', 'E501', 'W503', 'B950']
per-file-ignores = [
    '__init__.py:F401',
]
max-line-length = 90
max-complexity = 12
select = ['B', 'C', 'E', 'F', 'W', 'B9']

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
